--- app/models/voucher.rb
@@ -1,9 +1,0 @@
-class Voucher < ApplicationRecord
-  belongs_to :retailer
-  has_many :voucher_targets
-  belongs_to :item
-  has_one :product, through: :item
-
-  enum status: [ :active, :terminated , :canceled ]
-  
-end
