--- app/models/voucher_target.rb
@@ -1,6 +1,0 @@
-class VoucherTarget < ApplicationRecord
-  belongs_to :user
-  belongs_to :voucher
-
-  validates :user_id, uniqueness: { scope: :voucher_id }
-end
