+++ app/models/vouchers/voucher.rb
@@ -1,0 +1,11 @@
+module Vouchers
+  class Voucher < ApplicationRecord
+    belongs_to :retailer
+    has_many :voucher_targets
+    belongs_to :item
+    has_one :product, through: :item
+  
+    enum status: [ :active, :terminated , :canceled ]
+    
+  end
+end
