+++ app/models/vouchers/voucher_target.rb
@@ -1,0 +1,8 @@
+module Vouchers
+  class VoucherTarget < ApplicationRecord
+    belongs_to :user
+    belongs_to :voucher
+  
+    validates :user_id, uniqueness: { scope: :voucher_id }
+  end
+end
